#981 fixing layout and adding comments
