+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/ClassicMenu/Commands/VsCommands.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Input;
 using Windows.UI.Popups;
