+++ Microsoft.Toolkit.Uwp.UI.Controls/ClassicMenu/ClassicMenu.cs
@@ -1,12 +1,12 @@
 ﻿using System.Collections.Generic;
-using Windows.System;
-using Windows.UI.Core;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Input;
 
 namespace Microsoft.Toolkit.Uwp.UI.Controls
 {
+    /// <summary>
+    /// Classic Menu Control defines a menu of choices for users to invoke.
+    /// </summary>
     public partial class ClassicMenu : ItemsControl
     {
         private const string CtrlValue = "CTRL";
@@ -15,6 +15,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         // even if we have multiple menus in the same page we need only one cache because only one menu item will have certain short cut.
         private static readonly Dictionary<string, MenuFlyoutItem> MenuItemInputGestureCache;
 
+        /// <summary>
+        /// Gets or sets the orientation of the Classic menu, Horizontal or vertical means that child controls will be added horizontally until the width of the panel can't fit more control then a new row is added to fit new horizontal added child controls, vertical means that child will be added vertically until the height of the panel is recieved then a new column is added
+        /// </summary>
         public Orientation Orientation
         {
             get { return (Orientation)GetValue(OrientationProperty); }
@@ -37,6 +40,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
             new Dictionary<string, MenuFlyoutItem>();
         }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ClassicMenu"/> class.
+        /// </summary>
         public ClassicMenu()
         {
             DefaultStyleKey = typeof(ClassicMenu);
@@ -48,49 +54,6 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
             Loaded -= ClassicMenu_Loaded;
             Loaded += ClassicMenu_Loaded;
             base.OnApplyTemplate();
-        }
-
-        private void ClassicMenu_Loaded(object sender, RoutedEventArgs e)
-        {
-            var wrapPanel = (WrapPanel.WrapPanel)ItemsPanelRoot;
-            wrapPanel.Orientation = Orientation;
-
-            Dispatcher.AcceleratorKeyActivated += Dispatcher_AcceleratorKeyActivated;
-            Window.Current.CoreWindow.KeyDown += CoreWindow_KeyDown;
-        }
-
-        private void CoreWindow_KeyDown(CoreWindow sender, KeyEventArgs args)
-        {
-            var element = FocusManager.GetFocusedElement();
-            var menuItem = element as ClassicMenuItem;
-            if (menuItem != null)
-            {
-                if (NavigateThrowMenuHeader(args, this, menuItem, Orientation))
-                {
-                    return;
-                }
-            }
-
-            string gestureKey = MapInputToGestureKey(args);
-
-            if (gestureKey == null)
-            {
-                return;
-            }
-
-            if (MenuItemInputGestureCache.ContainsKey(gestureKey))
-            {
-                var cachedMenuItem = MenuItemInputGestureCache[gestureKey];
-                cachedMenuItem.Command?.Execute(cachedMenuItem.CommandParameter);
-            }
-        }
-
-        private void Dispatcher_AcceleratorKeyActivated(CoreDispatcher sender, AcceleratorKeyEventArgs args)
-        {
-            if (args.VirtualKey == VirtualKey.Menu && !args.KeyStatus.WasKeyDown)
-            {
-                Focus(FocusState.Keyboard);
-            }
         }
     }
 }
