+++ Microsoft.Toolkit.Uwp.UI.Controls/ClassicMenu/ClassicMenuItem.cs
@@ -5,9 +5,18 @@ using Windows.UI.Xaml.Controls.Primitives;
 
 namespace Microsoft.Toolkit.Uwp.UI.Controls
 {
+    /// <summary>
+    /// Classic Menu Item is the items main container for Class Menu control
+    /// </summary>
     public class ClassicMenuItem : ItemsControl
     {
+        /// <summary>
+        /// ClassMenuItem header text
+        /// </summary>
         public static readonly DependencyProperty HeaderProperty = DependencyProperty.Register(nameof(Header), typeof(string), typeof(ClassicMenuItem), new PropertyMetadata(default(string)));
+
+        private ClassicMenu ParentMenu => this.FindAscendant<ClassicMenu>();
+
         private Button _flyoutButton;
 
         /// <summary>
@@ -19,12 +28,18 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
             set { SetValue(HeaderProperty, value); }
         }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ClassicMenuItem"/> class.
+        /// </summary>
         public ClassicMenuItem()
         {
             DefaultStyleKey = typeof(ClassicMenuItem);
             IsFocusEngagementEnabled = true;
         }
 
+        /// <summary>
+        /// this method is used to show the menu for current item
+        /// </summary>
         public void ShowMenu()
         {
             _flyoutButton?.Flyout?.ShowAt(_flyoutButton);
@@ -56,7 +71,5 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
 
             base.OnApplyTemplate();
         }
-        
-        private ClassicMenu ParentMenu => this.FindAscendant<ClassicMenu>();
     }
 }
